Add optimize/queue-assets for batching individual assets
We ran into an issue when we were switching to this plugin where we have
16K images that needed processing. The one task to process all images
took about a week because we had to restart it multiple times and
manually change the offset to make sure it would start at new images.

It didn't occur until we were about to launch that there already
existed a queue job that would process a single image. So we wrote a
quick script to batch those.

I wasn't sure if this was the type of thing that a plugin would want to
support, but wanted to offer it in the event you've had other devs run
into a similar problem.